style: Apply Prettier code formatting
