+++ src/routes/index.ts
@@ -1,0 +1,6 @@
+import { Express } from "express";
+import { Auth } from "./auth";
+
+export const routes = (app: Express) => {
+	app.use("/auth", Auth());
+};
