+++ src/util/asyncWrapper.ts
@@ -3,9 +3,9 @@ import { Request, Response, NextFunction } from "express";
 export const asyncWrapper = (fn: Function) => {
 	return async (req: Request, res: Response, next: NextFunction) => {
 		try {
-			await fn(req, res, next)
+			await fn(req, res, next);
 		} catch (error) {
-			next(error)
+			next(error);
 		}
-	}
-}
+	};
+};
