+++ tsconfig.json
@@ -1,11 +1,11 @@
 {
   "compilerOptions": {
-    "target": "ES2021", // 어떤 버전으로 컴파일할지 작성 
+    "target": "ES2021", // 어떤 버전으로 컴파일할지 작성
     "module": "commonjs", // 어떤 모듈 방식으로 컴파일할지 설정
-    "outDir": "./build/src",	//컴파일 후 js 파일들이 생성되는 곳
+    "outDir": "./build/src", //컴파일 후 js 파일들이 생성되는 곳
     "esModuleInterop": true,
-    "rootDir": ".",	//루트 폴더
-    "strict": true,	//strict 옵션 활성화
-    "moduleResolution": "node",	//모듈 해석 방법 설정: 'node' (Node.js)
+    "rootDir": "./src", //루트 폴더
+    "strict": true, //strict 옵션 활성화
+    "moduleResolution": "node" //모듈 해석 방법 설정: 'node' (Node.js)
   }
 }
